+++ semcore/data-table/src/components/Body/Body.types.ts
@@ -1,5 +1,7 @@
-import type { Theme } from './Cell.types';
-import type { DTRow } from './Row.types';
+import type * as React from 'react';
+
+import type { CellPropsInner, Theme } from './Cell.types';
+import type { DTRow, UniqRowKey } from './Row.types';
 import type { DTUse, VirtualScroll } from '../DataTable/DataTable.types';
 import type { DTColumn } from '../Head/Column.types';
 
@@ -53,7 +55,7 @@ export type BodyPropsInner = DataTableBodyProps & {
   renderCell?: (props: CellRenderProps) => React.ReactNode | Record<string, any>;
   onBackFromAccordion: (colIndex: number) => void;
   stickyHeader?: boolean;
-  selectedRows?: number[];
+  selectedRows?: UniqRowKey[];
   onSelectRow?: (
     isSelect: boolean,
     selectedRowIndex: number,
@@ -66,4 +68,5 @@ export type BodyPropsInner = DataTableBodyProps & {
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
   accordionDuration?: number | [number, number];
+  onCellClick: CellPropsInner['onClick'];
 };