+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -230,12 +230,15 @@ class DataTableRoot<D extends DataTableData> extends Component<
       totalRows: this.totalRows,
       selectedRows: selectedRows,
       onChangeSelectAll: (value, e) => {
-        const selectedRowsIndexes = value
-          ? new Array(this.totalRows).fill(undefined).map((_, i) => i)
-          : [];
-        onSelectedRowsChange?.(selectedRowsIndexes, e);
+        if (value === false) {
+          onSelectedRowsChange?.([], e);
+        } else {
+          const selectedRows = this.flatRows.map((row) => row[UNIQ_ROW_KEY]);
+          onSelectedRowsChange?.(selectedRows, e);
+        }
       },
       getFixedStyle: this.getFixedStyle,
+      onCellClick: this.handleCellClick,
       ...headerProps,
     };
   }
@@ -290,8 +293,16 @@ class DataTableRoot<D extends DataTableData> extends Component<
       selectedRows,
       onSelectRow: this.handleSelectRow,
       getFixedStyle: this.getFixedStyle,
+      onCellClick: this.handleCellClick,
     };
   }
+
+  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow }) => {
+    if (lastInteraction.isMouse()) {
+      console.log('click on cell in data table handler', opt);
+      this.initFocusableCell([this.hasFocusableInHeader() ? opt.rowIndex + 1 : opt.rowIndex, opt.colIndex]);
+    }
+  };
 
   handleSelectRow = (
     isSelected: boolean,
@@ -304,10 +315,10 @@ class DataTableRoot<D extends DataTableData> extends Component<
     if (selectedRows && onSelectedRowsChange) {
       const newSelectedRows = new Set(selectedRows);
 
-      if (isSelected && !newSelectedRows.has(selectedRowIndex)) {
-        newSelectedRows.add(selectedRowIndex);
-      } else if (!isSelected && newSelectedRows.has(selectedRowIndex)) {
-        newSelectedRows.delete(selectedRowIndex);
+      if (isSelected && !newSelectedRows.has(row[UNIQ_ROW_KEY])) {
+        newSelectedRows.add(row[UNIQ_ROW_KEY]);
+      } else if (!isSelected && newSelectedRows.has(row[UNIQ_ROW_KEY])) {
+        newSelectedRows.delete(row[UNIQ_ROW_KEY]);
       }
 
       onSelectedRowsChange([...newSelectedRows], event, {
@@ -357,7 +368,8 @@ class DataTableRoot<D extends DataTableData> extends Component<
   };
 
   hasFocusableInHeader = () => {
-    return this.headerRef.current && hasFocusableIn(this.headerRef.current);
+    return (this.headerRef.current && hasFocusableIn(this.headerRef.current)) ||
+      this.columns.some((column) => column.sortable);
   };
 
   onExpandRow = (expandedRow: DTRow) => {
@@ -519,13 +531,18 @@ class DataTableRoot<D extends DataTableData> extends Component<
     }
   };
 
-  initFocusableCell = () => {
+  initFocusableCell(): void;
+  initFocusableCell(initCell: [row: number, cell: number]): void;
+  initFocusableCell(initCell?: [row: number, cell: number]) {
     const hasFocusable = this.hasFocusableInHeader();
 
+    const initRow = initCell?.[0] ?? 0;
+    const initCol = initCell?.[1] ?? 0;
+
     if (hasFocusable) {
-      this.focusedCell = [0, 0];
+      this.focusedCell = [initRow, initCol];
     } else {
-      this.focusedCell = [1, 0];
+      this.focusedCell = [initRow + 1, initCol];
     }
   };
 
@@ -650,7 +667,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     }
 
     let scrollDirection: 'both' | 'horizontal' | 'vertical' | undefined = undefined;
-    const hasWidthSettings = (Boolean(w) && w !== '100%') || Boolean(wMax);
+    const hasWidthSettings = Boolean(w) || Boolean(wMax);
     const hasHeightSettings = (Boolean(h) && h !== 'fit-content') || Boolean(hMax);
 
     if (hasWidthSettings && !hasHeightSettings) {
@@ -1056,7 +1073,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
   private calculateRows(): Array<DTRow[] | DTRow> {
     const columns = this.columns;
     // @ts-ignore
-    const { data, uid } = this.props;
+    const { data, uid, uniqueRowKey } = this.props;
 
     const rows: Array<DTRow[] | DTRow> = [];
     const columnNames = columns.map((column: DTColumn) => column.name);
@@ -1092,7 +1109,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
         },
         {
           [UNIQ_ROW_KEY]:
-            row[UNIQ_ROW_KEY] || (`${uid}_${(rowIndex + id).toString(36)}` as UniqRowKey),
+            row[UNIQ_ROW_KEY] || (uniqueRowKey ? row[uniqueRowKey] : `${uid}_${(rowIndex + id).toString(36)}`) as UniqRowKey,
           [ROW_INDEX]: rowIndex,
         },
       );
@@ -1179,7 +1196,16 @@ class DataTableRoot<D extends DataTableData> extends Component<
     let height = 0;
 
     for (let i = 0; i < (header?.length ?? 0); i++) {
-      const columnHeight = header?.item(i)?.getBoundingClientRect().height;
+      const item = header?.item(i);
+      let columnHeight = item?.getBoundingClientRect().height;
+
+      if (item instanceof HTMLElement && item.dataset.groupContainer) {
+        const groupHeight = item.children.item(0)?.getBoundingClientRect().height ?? 0;
+        const cellHeight = item.children.item(1)?.getBoundingClientRect().height ?? 0;
+
+        columnHeight = groupHeight + cellHeight;
+      }
+
       if (columnHeight) {
         height = columnHeight;
         break;