+++ semcore/data-table/src/components/Head/Head.types.ts
@@ -1,4 +1,6 @@
 import type { DTColumn } from './Column.types';
+import type { CellPropsInner } from '../Body/Cell.types';
+import type { UniqRowKey } from '../Body/Row.types';
 import type { DataTableData, DataTableProps, DTUse } from '../DataTable/DataTable.types';
 
 export type DataTableHeadProps = {
@@ -40,10 +42,11 @@ export type HeadPropsInner<D extends DataTableData> = {
   sideIndents?: 'wide';
 
   totalRows: number;
-  selectedRows?: number[];
+  selectedRows?: UniqRowKey[];
   onChangeSelectAll?: (value: boolean, event?: React.SyntheticEvent<HTMLElement>) => void;
 
   getFixedStyle: (
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
+  onCellClick: CellPropsInner['onClick'];
 };