+++ semcore/select/__tests__/select.browser-test.tsx
@@ -306,6 +306,15 @@ test.describe('Basic select', () => {
       await expect(menu).not.toBeVisible();
     });
 
+    await test.step('Verify 1st item highlighted when interacting by mouse and the by keyboard', async () => {
+      await select.click();
+      await page.keyboard.press('ArrowDown');
+      await expect(menu).toBeVisible();
+      await expect(options.first()).toHaveClass(/highlighted/);
+      await page.keyboard.press('Escape');
+      await expect(menu).not.toBeVisible();
+    });
+
     await test.step('Verify opens by Enter', async () => {
       await page.keyboard.press('Enter');
       await expect(menu).toBeVisible();
@@ -327,29 +336,29 @@ test.describe('Basic select', () => {
     await test.step('Verify opens by ArrowUp', async () => {
       await page.keyboard.press('ArrowUp');
       await expect(menu).toBeVisible();
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.first()).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify Arrows navigation in menu', async () => {
       await page.keyboard.press('ArrowRight');
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.nth(0)).toHaveClass(/highlighted/);
       await page.keyboard.press('ArrowLeft');
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.nth(0)).toHaveClass(/highlighted/);
 
       await page.keyboard.press('ArrowDown');
-      await expect(options.nth(0)).toHaveClass(/highlighted/);
+      await expect(options.nth(1)).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify Tab not removed focus outside menu', async () => {
       await page.keyboard.press('Tab');
-      await expect(options.nth(0)).toHaveClass(/highlighted/);
+      await expect(options.nth(1)).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify space selects item and closes menu', async () => {
       await page.keyboard.press('Space');
       await expect(menu).not.toBeVisible();
-      await expect(select).toHaveAttribute('value', '0');
-      await expect(triggerText).toHaveText(/Option 0/);
+      await expect(select).toHaveAttribute('value', '1');
+      await expect(triggerText).toHaveText(/Option 1/);
       await expect(select).toBeFocused();
     });
 
@@ -359,8 +368,8 @@ test.describe('Basic select', () => {
       await page.keyboard.press('ArrowDown');
       await page.keyboard.press('Enter');
       await expect(menu).not.toBeVisible();
-      await expect(select).toHaveAttribute('value', '1');
-      await expect(triggerText).toHaveText(/Option 1/);
+      await expect(select).toHaveAttribute('value', '2');
+      await expect(triggerText).toHaveText(/Option 2/);
       await expect(select).toBeFocused();
     });
 
@@ -369,8 +378,8 @@ test.describe('Basic select', () => {
       await new Promise((resolve) => setTimeout(resolve, 500));
       await page.keyboard.press('Escape');
       await expect(menu).not.toBeVisible();
-      await expect(select).toHaveAttribute('value', '1');
-      await expect(triggerText).toHaveText(/Option 1/);
+      await expect(select).toHaveAttribute('value', '2');
+      await expect(triggerText).toHaveText(/Option 2/);
       await expect(select).toBeFocused();
     });
   });
@@ -461,29 +470,29 @@ test.describe('Basic select', () => {
     await test.step('Verify opens by ArrowUp', async () => {
       await page.keyboard.press('ArrowUp');
       await expect(menu).toBeVisible();
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.first()).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify Arrows navigation in menu', async () => {
       await page.keyboard.press('ArrowRight');
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.first()).toHaveClass(/highlighted/);
       await page.keyboard.press('ArrowLeft');
-      await expect(options.nth(5)).toHaveClass(/highlighted/);
+      await expect(options.first()).toHaveClass(/highlighted/);
 
       await page.keyboard.press('ArrowDown');
-      await expect(options.nth(0)).toHaveClass(/highlighted/);
+      await expect(options.nth(1)).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify Tab not removed focus outside menu', async () => {
       await page.keyboard.press('Tab');
-      await expect(options.nth(0)).toHaveClass(/highlighted/);
+      await expect(options.nth(1)).toHaveClass(/highlighted/);
     });
 
     await test.step('Verify space selects item and closes menu', async () => {
       await page.keyboard.press('Space');
       await expect(menu).not.toBeVisible();
-      await expect(select).toHaveAttribute('value', '0');
-      await expect(triggerText).toHaveText(/Label 0/);
+      await expect(select).toHaveAttribute('value', '1');
+      await expect(triggerText).toHaveText(/Label 1/);
       await expect(select).toBeFocused();
     });
 
@@ -493,8 +502,8 @@ test.describe('Basic select', () => {
       await page.keyboard.press('ArrowDown');
       await page.keyboard.press('Enter');
       await expect(menu).not.toBeVisible();
-      await expect(select).toHaveAttribute('value', '1');
-      await expect(triggerText).toHaveText(/Label 1/);
+      await expect(select).toHaveAttribute('value', '2');
+      await expect(triggerText).toHaveText(/Label 2/);
       await expect(select).toBeFocused();
     });
   });
@@ -761,7 +770,7 @@ test.describe('Options filtering', () => {
     });
   });
 
-  test('Verify advances filterring control base actions', async ({ page }) => {
+  test('Verify advances filtering control base actions', async ({ page }) => {
     const standPath = 'stories/components/select/docs/examples/advanced_filtering_control.tsx';
     const htmlContent = await e2eStandToHtml(standPath, 'en');
 
@@ -855,5 +864,59 @@ test.describe('Focus interaction', () => {
     await errorTooltip.waitFor();
 
     await expect(errorTooltip).toBeVisible();
+  });
+});
+
+test.describe('Sticky groups', () => {
+  test('Verify sticky groups mouse scroll', async ({ page, browserName }) => {
+    const standPath = 'stories/components/select/docs/examples/sticky_groups.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    const { select, menu, options } = getSelectLocators(page);
+
+    await page.locator('[data-ui-name="Select.Trigger"]').click();
+    await options.first().waitFor({ state: 'visible' });
+
+    await options.first().hover();
+    await page.mouse.wheel(0, 2000);
+    await page.waitForTimeout(1000);
+    await expect(page).toHaveScreenshot({ maxDiffPixelRatio: 0.01 });
+  });
+
+  test('Verify sticky groups keyboard scroll', async ({ page, browserName }) => {
+    const standPath = 'stories/components/select/docs/examples/sticky_groups.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+    const { select, menu, options } = getSelectLocators(page);
+
+    await page.keyboard.press('Tab');
+    await page.keyboard.press('Enter');
+
+    await expect(page.locator('[data-ui-name="Select.InputSearch"]')).toBeFocused();
+    await page.keyboard.press('ArrowDown');
+    await expect(options.nth(1)).not.toHaveClass(/highlighted/);
+
+    for (let i = 0; i < 14; i++) {
+      await page.keyboard.press('ArrowDown');
+      await page.waitForTimeout(100);
+    }
+    await expect(options.nth(14)).toHaveClass(/highlighted/);
+
+    await page.keyboard.press('Enter');
+    await expect(options.nth(14)).not.toBeVisible();
+
+    await page.keyboard.press('Enter');
+    await options.nth(14).waitFor({ state: 'visible' });
+
+    await expect(page).toHaveScreenshot({ maxDiffPixelRatio: 0.01 });
+    // too unstable, fix it later
+    // if (browserName !== 'chromium') return;
+    // await expect(options.nth(14)).toHaveClass(/highlighted/);
+    // await page.keyboard.press('Tab');
+    // await page.keyboard.press('Tab');
+    //
+    // await expect(page.locator('[data-ui-name="Select.InputSearch"]')).toBeFocused();
   });
 });